Extract sheet modifiers and clarify naming
